hydra-queue-runner: Fix use of pow() and round()
Functions like powf() and roundf() are now longer available without
inclusion of <cmath> since GCC 5.4 and so the build of Hydra fails.

I presume this was a bug in older versions of GCC and these functions
shouldn't be exposed (I'm guessing they were exposed by some of the
includes used in builder.cc, but I haven't checked).

Unfortunately I haven't found anything about such a bug in the upstream
changelog or the bugtracker.

Nevertheless, we shouldn't use powf() and roundf() in any case, because
they're C99 functions:

http://en.cppreference.com/w/c/numeric/math/pow

C++11 (which we explicitly enforce using -std=c++11) on the other side
doesn't have these functions at all:

http://en.cppreference.com/w/cpp/numeric/math

So I *think* these functions shouldn't even be exposed via <cmath> and
that's probably something where GCC will catch up in future versions.

To be safe in any case, I'm now using the generalized std::pow() and
std::round() functions.

+++ src/hydra-queue-runner/builder.cc
@@ -38,7 +38,7 @@ void State::builder(MachineReservation::ptr reservation)
             step_->tries++;
             nrRetries++;
             if (step_->tries > maxNrRetries) maxNrRetries = step_->tries; // yeah yeah, not atomic
-            int delta = retryInterval * powf(retryBackoff, step_->tries - 1) + (rand() % 10);
+            int delta = retryInterval * std::pow(retryBackoff, step_->tries - 1) + (rand() % 10);
             printMsg(lvlInfo, format("will retry ‘%1%’ after %2%s") % step->drvPath % delta);
             step_->after = std::chrono::system_clock::now() + std::chrono::seconds(delta);
         }
